+++ packages/dialect-react/components/DialectContext/index.tsx
@@ -56,6 +56,7 @@ const swrFetchMetadata = (
 type PropsType = {
   children: JSX.Element;
   publicKey?: anchor.web3.PublicKey;
+  pollingInterval?: number;
 };
 
 // TODO: revisit api functions and errors to be moved out from context
@@ -131,6 +132,8 @@ export const DialectProvider = (props: PropsType): JSX.Element => {
   const [encryptionProps, setEncryptionProps] =
     React.useState<EncryptionProps | null>(null);
 
+  const pollingInterval = props.pollingInterval || POLLING_INTERVAL_MS;
+
   const getEncryptionProps =
     useCallback(async (): Promise<EncryptionProps | null> => {
       if (!wallet || isAnchorWallet(wallet) || walletName !== 'Sollet') {
@@ -182,7 +185,7 @@ export const DialectProvider = (props: PropsType): JSX.Element => {
     false && wallet && program ? ['metadata', program] : null,
     swrFetchMetadata,
     {
-      refreshInterval: POLLING_INTERVAL_MS,
+      refreshInterval: pollingInterval,
     }
   );
 
@@ -196,7 +199,7 @@ export const DialectProvider = (props: PropsType): JSX.Element => {
       : null,
     swrFetchDialects,
     {
-      refreshInterval: POLLING_INTERVAL_MS,
+      refreshInterval: pollingInterval,
     }
   );
 
@@ -221,7 +224,7 @@ export const DialectProvider = (props: PropsType): JSX.Element => {
       : null,
     swrFetchDialect,
     {
-      refreshInterval: POLLING_INTERVAL_MS,
+      refreshInterval: pollingInterval,
     }
   );
 
@@ -231,9 +234,8 @@ export const DialectProvider = (props: PropsType): JSX.Element => {
         program as anchor.Program,
         props.publicKey
       ).then(([address, _]: [anchor.web3.PublicKey, number]) => {
-            setDialectAddress(address.toBase58())
-          }
-      );
+        setDialectAddress(address.toBase58());
+      });
     }
     return;
   }, [program, props.publicKey]);