+++ packages/dialect-react-ui/CHANGELOG.md
@@ -15,6 +15,7 @@
 ## [0.8.7] - 2022-06-01
 
 - Add solana name service support
+- Add polling props to notfication button
 
 ## [0.8.6] - 2022-05-11