+++ packages/dialect-react/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [UNRELEASED]
 
+- added polling interval to context. Can be changed from UI
+
 ## [0.5.4] - 2022-05-11
 
 - fix: allow to pass wallet adapter to ApiContext